# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "CSE 341 Project 2 - Fitness Tracker API",
        "description": "REST API for tracking workouts and managing users in a fitness application",
        "version": "1.0.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
        "/workouts/": {
            "get": {
                "description": "",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "userId": {
                                    "example": "any"
                                },
                                "workoutName": {
                                    "example": "any"
                                },
                                "date": {
                                    "example": "any"
                                },
                                "duration": {
                                    "example": "any"
                                },
                                "caloriesBurned": {
                                    "example": "any"
                                },
                                "exerciseType": {
                                    "example": "any"
                                },
                                "notes": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/workouts/{id}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "userId": {
                                    "example": "any"
                                },
                                "workoutName": {
                                    "example": "any"
                                },
                                "date": {
                                    "example": "any"
                                },
                                "duration": {
                                    "example": "any"
                                },
                                "caloriesBurned": {
                                    "example": "any"
                                },
                                "exerciseType": {
                                    "example": "any"
                                },
                                "notes": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/": {
            "get": {
                "description": "",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "example": "any"
                                },
                                "email": {
                                    "example": "any"
                                },
                                "age": {
                                    "example": "any"
                                },
                                "weight": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "example": "any"
                                },
                                "email": {
                                    "example": "any"
                                },
                                "age": {
                                    "example": "any"
                                },
                                "weight": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/": {
            "get": {
                "description": "",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "6830c1332164274e3bf8b9b3"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "age": {
                    "type": "number",
                    "example": 25
                },
                "weight": {
                    "type": "number",
                    "example": 180
                },
                "createdAt": {
                    "type": "string",
                    "example": "2024-06-03T10:30:00.000Z"
                }
            }
        },
        "UserInput": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "age": {
                    "type": "number",
                    "example": 25
                },
                "weight": {
                    "type": "number",
                    "example": 180
                }
            }
        },
        "Workout": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "6830c1332164274e3bf8b9b4"
                },
                "userId": {
                    "type": "string",
                    "example": "6830c1332164274e3bf8b9b3"
                },
                "workoutName": {
                    "type": "string",
                    "example": "Morning Run"
                },
                "date": {
                    "type": "string",
                    "example": "2024-06-03T00:00:00.000Z"
                },
                "duration": {
                    "type": "number",
                    "example": 30
                },
                "caloriesBurned": {
                    "type": "number",
                    "example": 300
                },
                "exerciseType": {
                    "type": "string",
                    "example": "cardio"
                },
                "notes": {
                    "type": "string",
                    "example": "Beautiful morning run in the park"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2024-06-03T10:30:00.000Z"
                }
            }
        },
        "WorkoutInput": {
            "type": "object",
            "properties": {
                "userId": {
                    "type": "string",
                    "example": "6830c1332164274e3bf8b9b3"
                },
                "workoutName": {
                    "type": "string",
                    "example": "Morning Run"
                },
                "date": {
                    "type": "string",
                    "example": "2024-06-03"
                },
                "duration": {
                    "type": "number",
                    "example": 30
                },
                "caloriesBurned": {
                    "type": "number",
                    "example": 300
                },
                "exerciseType": {
                    "type": "string",
                    "example": "cardio"
                },
                "notes": {
                    "type": "string",
                    "example": "Beautiful morning run in the park"
                }
            }
        }
    }
}